# Conflicts:
#	src/backEnd/node_modules/.bin/color-support
#	src/backEnd/node_modules/.bin/detect-libc
#	src/backEnd/node_modules/.bin/mkdirp
#	src/backEnd/node_modules/.bin/node-pre-gyp
#	src/backEnd/node_modules/.bin/nopt
#	src/backEnd/node_modules/.bin/rimraf
#	src/backEnd/node_modules/.bin/semver
#	src/backEnd/node_modules/@mapbox/node-pre-gyp/bin/node-pre-gyp
#	src/backEnd/node_modules/@mapbox/node-pre-gyp/package.json
#	src/backEnd/node_modules/abbrev/package.json
#	src/backEnd/node_modules/agent-base/node_modules/debug/package.json
#	src/backEnd/node_modules/agent-base/node_modules/ms/package.json
#	src/backEnd/node_modules/agent-base/package.json
#	src/backEnd/node_modules/ansi-regex/package.json
#	src/backEnd/node_modules/aproba/package.json
#	src/backEnd/node_modules/are-we-there-yet/node_modules/readable-stream/package.json
#	src/backEnd/node_modules/are-we-there-yet/package.json
#	src/backEnd/node_modules/balanced-match/package.json
#	src/backEnd/node_modules/bcrypt/appveyor.yml
#	src/backEnd/node_modules/bcrypt/lib/binding/napi-v3/bcrypt_lib.node
#	src/backEnd/node_modules/bcrypt/package.json
#	src/backEnd/node_modules/bcrypt/test_alpine.sh
#	src/backEnd/node_modules/brace-expansion/package.json
#	src/backEnd/node_modules/chownr/package.json
#	src/backEnd/node_modules/color-support/bin.js
#	src/backEnd/node_modules/color-support/package.json
#	src/backEnd/node_modules/concat-map/package.json
#	src/backEnd/node_modules/console-control-strings/package.json
#	src/backEnd/node_modules/cors/package.json
#	src/backEnd/node_modules/delegates/package.json
#	src/backEnd/node_modules/detect-libc/bin/detect-libc.js
#	src/backEnd/node_modules/detect-libc/package.json
#	src/backEnd/node_modules/emoji-regex/package.json
#	src/backEnd/node_modules/fs-minipass/package.json
#	src/backEnd/node_modules/fs.realpath/package.json
#	src/backEnd/node_modules/gauge/package.json
#	src/backEnd/node_modules/glob/package.json
#	src/backEnd/node_modules/has-unicode/package.json
#	src/backEnd/node_modules/https-proxy-agent/node_modules/debug/package.json
#	src/backEnd/node_modules/https-proxy-agent/node_modules/ms/package.json
#	src/backEnd/node_modules/https-proxy-agent/package.json
#	src/backEnd/node_modules/inflight/package.json
#	src/backEnd/node_modules/is-fullwidth-code-point/package.json
#	src/backEnd/node_modules/lru-cache/package.json
#	src/backEnd/node_modules/make-dir/node_modules/.bin/semver
#	src/backEnd/node_modules/make-dir/node_modules/semver/bin/semver.js
#	src/backEnd/node_modules/make-dir/node_modules/semver/package.json
#	src/backEnd/node_modules/make-dir/package.json
#	src/backEnd/node_modules/minimatch/package.json
#	src/backEnd/node_modules/minipass/package.json
#	src/backEnd/node_modules/minizlib/package.json
#	src/backEnd/node_modules/mkdirp/bin/cmd.js
#	src/backEnd/node_modules/mkdirp/package.json
#	src/backEnd/node_modules/node-addon-api/package.json
#	src/backEnd/node_modules/node-addon-api/tools/conversion.js
#	src/backEnd/node_modules/node-fetch/package.json
#	src/backEnd/node_modules/nopt/bin/nopt.js
#	src/backEnd/node_modules/nopt/package.json
#	src/backEnd/node_modules/npmlog/package.json
#	src/backEnd/node_modules/object-assign/package.json
#	src/backEnd/node_modules/once/package.json
#	src/backEnd/node_modules/path-is-absolute/package.json
#	src/backEnd/node_modules/rimraf/bin.js
#	src/backEnd/node_modules/rimraf/package.json
#	src/backEnd/node_modules/semver/bin/semver.js
#	src/backEnd/node_modules/semver/package.json
#	src/backEnd/node_modules/set-blocking/package.json
#	src/backEnd/node_modules/signal-exit/package.json
#	src/backEnd/node_modules/string-width/package.json
#	src/backEnd/node_modules/strip-ansi/package.json
#	src/backEnd/node_modules/tar/package.json
#	src/backEnd/node_modules/tr46/package.json
#	src/backEnd/node_modules/webidl-conversions/package.json
#	src/backEnd/node_modules/whatwg-url/lib/url-state-machine.js
#	src/backEnd/node_modules/whatwg-url/package.json
#	src/backEnd/node_modules/wide-align/LICENSE
#	src/backEnd/node_modules/wide-align/README.md
#	src/backEnd/node_modules/wide-align/align.js
#	src/backEnd/node_modules/wide-align/package.json
#	src/backEnd/node_modules/wrappy/package.json
#	src/backEnd/node_modules/yallist/package.json

+++ src/backEnd/app.js
@@ -21,7 +21,7 @@ app.use(express.static('../frontEnd/the_actually_website'));
 var mysqlConnection = mysql.createConnection({
     host: 'localhost',
     user: 'root',
-    password: 'root',
+    password: 'Assassin990118',
     database: 'cmpt370'
 });
 
@@ -34,6 +34,8 @@ mysqlConnection.connect((err) => {
         console.log('DB connection failed \n Error: '+ JSON.stringify(err,undefined,2));
 });
 
+// var for which room is selected
+var selectedRoom = 0;
 
 // a function to check whether input (checkbox) is checked
 function parserInt(input) {
@@ -156,18 +158,28 @@ app.post('/read', (req, res) => {
 
 // Delete rooms
 app.post('/delRoom', function(req,res){
-    var name = req.body.name;
-    let sql = 'DELETE FROM roominfo WHERE name=?';
+    var room_name = req.body.room_name;
+    var room_manager = req.body.room_manager;
+    let sql = 'SELECT manager FROM roominfo WHERE name=?';
 
-    console.log(name);
+    // console.log(room_name);
+    // console.log(room_manager);
 
-    mysqlConnection.query(sql,[name], (err, result) => {
+    mysqlConnection.query(sql, room_name, (err, result) => {
         if(err) throw err;
-        console.log(result);
-
+        else if (room_manager == result[0].manager) {
+            let sql1 = 'DELETE FROM roominfo WHERE name=?';
+            mysqlConnection.query(sql1, [room_name], (err, result) => {
+                if(err) throw err;
+                // console.log(result);
+                res.send("all done.");
+            });
+            // res.send("all done.")
+        }
+        else res.send("The room name and room manager do not match.")
+        // console.log(sql);
+        // console.log(result[0].manager);
     });
-
-    res.send("all done");
 })
 
 
@@ -241,29 +253,33 @@ app.post('/login',(req, res) =>{
 });
 
 // add comments and rating
-app.post('/addComment',(req,res) =>{
-    var room_id = req.body.room_id;
-    var comment = req.body.comment;
-    var rating = req.body.rating;
-    var post = {room_id: room_id, comment: comment, rating: rating};
-    console.log(post);
+app.post('/postComment',(req,res) =>{
+    var room_id = selectedRoom;
+    if (room_id == 0){
+        res.send("no room")
+    } else {
+        var comment = req.body.comment;
+        var rate = req.body.rate;
+        var post = {room_id: room_id, comment: comment, rate: rate};
+        console.log(post);
 
 
-    mysqlConnection.query('INSERT INTO comments (room_id,comment,rating) VALUES (?,?,?)',[room_id,comment,rating], (err, result) => {
-        if(err) throw err;
-        console.log(result);
-        console.log('added');
-    });
-    res.send("all done");
+        mysqlConnection.query('INSERT INTO comments (room_id,comment,rating) VALUES (?,?,?)',[room_id,comment,rate], (err, result) => {
+            if(err) throw err;
+            console.log(result);
+            console.log('added');
+        });
+        res.send("all done");
+    }  
 });
 
 
 // Get comments and ratings for a room
 app.get('/getComments',function(req,res){
-    var room_id = req.body.room_id;
+    var room_id = selectedRoom;
 
 
-    mysqlConnection.query('SELECT * FROM comments WHERE room_id = ?',room_id, (err, result) => {
+    mysqlConnection.query('SELECT * FROM comments WHERE room_id=?',room_id, (err, result) => {
         if(err) throw err;
         console.log(result);
         res.send(result);
@@ -295,6 +311,14 @@ app.post('/addCustomer',(req,res) =>{
     });
     res.send("all done");
 });
+
+// route for getting the room id of selected room
+app.post('/selectRoom', (req, res) => {
+    //console.log(req.body);
+    selectedRoom = req.body.selectedRoom;
+    res.send("room "+selectedRoom+" is selected!");
+});
+
